Organize data and response models
