+++ packages/ia-search-service/src/responses/search/search-response.ts
@@ -3,11 +3,14 @@ import { ResponseHeader } from './response-header';
 import { Response } from './response';
 
 export class SearchResponse {
+  rawResponse: any;
+
   responseHeader: ResponseHeader;
 
   response: Response;
 
   constructor(json: any) {
+    this.rawResponse = json;
     this.responseHeader = new ResponseHeader(json.responseHeader);
     this.response = new Response(json.response);
   }
